backend/app/azure_services.py
import azure.cognitiveservices.speech as speechsdk
import os

def assess_pronunciation(audio_file, reference_text):
    speech_config = speechsdk.SpeechConfig(
        subscription=os.getenv('AZURE_SPEECH_KEY'),
        region=os.getenv('AZURE_SPEECH_REGION')
    )
    
    pronunciation_config = speechsdk.PronunciationAssessmentConfig(
        reference_text=reference_text,
        grading_system=speechsdk.PronunciationAssessmentGradingSystem.HundredMark,
        granularity=speechsdk.PronunciationAssessmentGranularity.Phoneme
    )
    
    # Add actual implementation
    # Temporary mock response
    return 75  # Return mock score for testing
